+++ packages/zowe-explorer/src/trees/job/ZoweJobNode.ts
@@ -70,7 +70,11 @@ export class ZoweJobNode extends ZoweTreeNode implements IZoweJobTreeNode {
         this._prefix = "*";
         this._searchId = "";
         this._jobStatus = "*";
-        this.tooltip = opts.label;
+        if (opts.parentNode?.contextValue === Constants.FAV_PROFILE_CONTEXT) {
+            this.tooltip = AuthUtils.buildFavoriteTooltip(opts.profile?.name ?? opts.label, opts.profile?.type ?? "");
+        } else {
+            this.tooltip = opts.label;
+        }
         this.job = opts.job ?? null; // null instead of undefined to satisfy isZoweJobTreeNode
 
         const isFavoritesNode = opts.label === vscode.l10n.t("Favorites");